django/mysite/requirements.txt
asgiref==3.8.1
Django==5.1.7
gunicorn==22.0.0
packaging==24.1
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
whitenoise==6.7.0
Pillow
pyOpenSSL
django-countries
django-localflavor
django-phonenumber-field[phonenumbers]
device_detector
requests
django-qr-code
# mfa
pyotp
qrcode
pillow
#
python-decouple
psycopg2-binary
#psycopg2
python-dateutil
# tenant
django-tenants
# dev
django-seed
faker
django-debug-toolbar
django-extensions
Werkzeug
django-debug-toolbar
# api
djangorestframework-simplejwt